+++ solidity/contracts/bancorx/BancorX.sol
@@ -2,11 +2,11 @@
 pragma solidity 0.6.12;
 
 import "@openzeppelin/contracts/math/SafeMath.sol";
+import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
 
 import "./interfaces/IBancorXUpgrader.sol";
 import "./interfaces/IBancorX.sol";
 import "../utility/ContractRegistryClient.sol";
-import "../utility/TokenHandler.sol";
 import "../utility/TokenHolder.sol";
 
 /**
@@ -19,8 +19,9 @@ import "../utility/TokenHolder.sol";
  * Reporting cross chain transfers works similar to standard multisig contracts, meaning that multiple
  * callers are required to report a transfer before tokens are released to the target account.
  */
-contract BancorX is IBancorX, TokenHandler, TokenHolder, ContractRegistryClient {
+contract BancorX is IBancorX, TokenHolder, ContractRegistryClient {
     using SafeMath for uint256;
+    using SafeERC20 for IERC20;
 
     // represents a transaction on another blockchain where tokens were destroyed/locked
     struct Transaction {
@@ -43,7 +44,7 @@ contract BancorX is IBancorX, TokenHandler, TokenHolder, ContractRegistryClient 
     uint256 public prevReleaseBlockNumber; // the block number of the last release transaction
     uint8 public minRequiredReports; // minimum number of required reports to release tokens
 
-    IERC20Token public override token; // erc20 token
+    IERC20 public override token; // erc20 token
 
     bool public xTransfersEnabled = true; // true if x transfers are enabled, false if not
     bool public reportingEnabled = true; // true if reporting is enabled, false if not
@@ -132,7 +133,7 @@ contract BancorX is IBancorX, TokenHandler, TokenHolder, ContractRegistryClient 
         uint256 _limitIncPerBlock,
         uint8 _minRequiredReports,
         IContractRegistry _registry,
-        IERC20Token _token
+        IERC20 _token
     )
         public
         ContractRegistryClient(_registry)
@@ -454,7 +455,8 @@ contract BancorX is IBancorX, TokenHandler, TokenHolder, ContractRegistryClient 
      * @param _amount  the amount of tokens to lock
      */
     function lockTokens(uint256 _amount) private {
-        safeTransferFrom(token, msg.sender, address(this), _amount);
+        token.safeTransferFrom(msg.sender, address(this), _amount);
+
         emit TokensLock(msg.sender, _amount);
     }
 
@@ -475,7 +477,7 @@ contract BancorX is IBancorX, TokenHandler, TokenHolder, ContractRegistryClient 
         prevReleaseBlockNumber = block.number;
 
         // no need to require, reverts on failure
-        safeTransfer(token, _to, _amount);
+        token.safeTransfer(_to, _amount);
 
         emit TokensRelease(_to, _amount);
     }