+++ solidity/contracts/converter/interfaces/IConverterRegistryData.sol
@@ -1,7 +1,9 @@
 // SPDX-License-Identifier: SEE LICENSE IN LICENSE
 pragma solidity 0.6.12;
+
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+
 import "./IConverterAnchor.sol";
-import "../../token/interfaces/IERC20Token.sol";
 
 interface IConverterRegistryData {
     function addSmartToken(IConverterAnchor _anchor) external;
@@ -12,9 +14,9 @@ interface IConverterRegistryData {
 
     function removeLiquidityPool(IConverterAnchor _liquidityPoolAnchor) external;
 
-    function addConvertibleToken(IERC20Token _convertibleToken, IConverterAnchor _anchor) external;
+    function addConvertibleToken(IERC20 _convertibleToken, IConverterAnchor _anchor) external;
 
-    function removeConvertibleToken(IERC20Token _convertibleToken, IConverterAnchor _anchor) external;
+    function removeConvertibleToken(IERC20 _convertibleToken, IConverterAnchor _anchor) external;
 
     function getSmartTokenCount() external view returns (uint256);
 
@@ -36,18 +38,18 @@ interface IConverterRegistryData {
 
     function getConvertibleTokens() external view returns (address[] memory);
 
-    function getConvertibleToken(uint256 _index) external view returns (IERC20Token);
+    function getConvertibleToken(uint256 _index) external view returns (IERC20);
 
     function isConvertibleToken(address _value) external view returns (bool);
 
-    function getConvertibleTokenSmartTokenCount(IERC20Token _convertibleToken) external view returns (uint256);
+    function getConvertibleTokenSmartTokenCount(IERC20 _convertibleToken) external view returns (uint256);
 
-    function getConvertibleTokenSmartTokens(IERC20Token _convertibleToken) external view returns (address[] memory);
+    function getConvertibleTokenSmartTokens(IERC20 _convertibleToken) external view returns (address[] memory);
 
-    function getConvertibleTokenSmartToken(IERC20Token _convertibleToken, uint256 _index)
+    function getConvertibleTokenSmartToken(IERC20 _convertibleToken, uint256 _index)
         external
         view
         returns (IConverterAnchor);
 
-    function isConvertibleTokenSmartToken(IERC20Token _convertibleToken, address _value) external view returns (bool);
+    function isConvertibleTokenSmartToken(IERC20 _convertibleToken, address _value) external view returns (bool);
 }