+++ solidity/contracts/utility/TokenHolder.sol
@@ -1,10 +1,12 @@
 // SPDX-License-Identifier: SEE LICENSE IN LICENSE
 pragma solidity 0.6.12;
+
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
+
 import "./Owned.sol";
 import "./Utils.sol";
-import "./TokenHandler.sol";
 import "./interfaces/ITokenHolder.sol";
-import "../token/interfaces/IERC20Token.sol";
 
 /**
  * @dev This contract provides a safety mechanism for allowing the owner to
@@ -17,7 +19,9 @@ import "../token/interfaces/IERC20Token.sol";
  * in order to support both non standard as well as standard token contracts.
  * see https://github.com/ethereum/solidity/issues/4116
  */
-contract TokenHolder is ITokenHolder, TokenHandler, Owned, Utils {
+contract TokenHolder is ITokenHolder, Owned, Utils {
+    using SafeERC20 for IERC20;
+
     /**
      * @dev withdraws tokens held by the contract and sends them to an account
      * can only be called by the owner
@@ -27,10 +31,10 @@ contract TokenHolder is ITokenHolder, TokenHandler, Owned, Utils {
      * @param _amount  amount to withdraw
      */
     function withdrawTokens(
-        IERC20Token _token,
+        IERC20 _token,
         address _to,
         uint256 _amount
     ) public virtual override ownerOnly validAddress(address(_token)) validAddress(_to) notThis(_to) {
-        safeTransfer(_token, _to, _amount);
+        _token.safeTransfer(_to, _amount);
     }
 }